+++ code/drasil-srs/lib/Drasil/SRS/DocumentLanguage/Units.hs
@@ -1,5 +1,5 @@
 -- | Helpers for dealing with formatting of Units
-module Drasil.DocumentLanguage.Units (toSentenceUnitless, toSentence) where
+module Drasil.SRS.DocumentLanguage.Units (toSentenceUnitless, toSentence) where
 
 import Language.Drasil (Sentence(S, Sy), usymb, MayHaveUnit(getUnit))
